app/requirements.txt
langchain==0.0.202
fastapi
openai
backoff
pathlib
openai
python-dotenv
wandb
uvicorn[standard]